project/requirements.txt
asgiref==3.2.3
Django==3.0
django-extra-fields==2.0.2
djangorestframework==3.10.3
Pillow==6.2.1
pytesseract==0.3.0
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
